+++ backend/src/main/java/tn/esprit/pi/tbibi/DTO/forumpost/ForumPostResponse.java
@@ -17,7 +17,7 @@ public class ForumPostResponse {
     LocalDateTime createdDate;
     Integer views;
     String postStatus;
-    Long authorId;
+    Integer authorId;
     String authorName;
     Long categoryId;
     String categoryName;
